day 5

+++ controller/videoController.js
@@ -30,17 +30,21 @@ export const postUpload = async (req, res) => {
 	const newVideo = await Video.create({
 		fileUrl     : path,
 		title,
-		description
+		description,
+		creator     : req.user.id
 	});
+	req.user.videos.push(newVideo.id);
+	req.user.save();
 	res.redirect(routes.videoDetail(newVideo.id));
 };
 
 export const videoDetail = async (req, res) => {
 	const { params: { id } } = req;
 	try {
-		const video = await Video.findById(id);
+		const video = await Video.findById(id).populate('creator');
 		res.render('videoDetail', { pageTitle: video.title, video });
 	} catch (error) {
+		console.log(error);
 		res.redirect(routes.root);
 	}
 };
@@ -49,7 +53,11 @@ export const getEditVideo = async (req, res) => {
 	const { params: { id } } = req;
 	try {
 		const video = await Video.findById(id);
-		res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+		if (video.creator !== req.user.id) {
+			throw Error();
+		} else {
+			res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+		}
 	} catch (error) {
 		res.redirect(routes.root);
 	}
@@ -73,7 +81,14 @@ export const deleteVideo = async (req, res) => {
 		console.log(`successfully deleted ${fileUrl}`);
 	});
 	try {
-		await Video.findOneAndDelete({ _id: id });
-	} catch (error) {}
+		const video = await Video.findById(id);
+		if (video.creator !== req.user.id) {
+			throw Error();
+		} else {
+			await Video.findOneAndDelete({ _id: id });
+		}
+	} catch (error) {
+		console.log(error);
+	}
 	res.redirect(routes.root);
 };
